2009-10-08      Chris Johns <chrisj@rtems.org>
        * libchip/ide/ata.c, c/src/libchip/ide/ata_internal.h: Remove the
        multiblock support from the ATA driver. Multiblock at the disk
        level should not be a global policy as a disk may have more than
        one partition with different block sizes. An IDE driver could
        decide to use the feature with DMA but this a driver specific
        design choice. Fixed bugs relating to variable block sizes and
        large block transfer.
        * c/src/libchip/ide/ide_controller.c,
        c/src/libchip/ide/ide_ctrl_cfg.h, c/src/libchip/ide/ide_ctrl_io.h:
        Changed the block size to 32bits so blocks of 64K or bigger can be
        transfered in a single driver call.

+++ c/src/libchip/ide/ide_controller.c
@@ -92,7 +92,7 @@ ide_controller_initialize(rtems_device_major_number  major,
  */
 void
 ide_controller_read_data_block(rtems_device_minor_number  minor,
-                               uint16_t                   block_size,
+                               uint32_t                   block_size,
                                rtems_blkdev_sg_buffer    *bufs,
                                uint32_t                  *cbuf,
                                uint32_t                  *pos)
@@ -121,7 +121,7 @@ ide_controller_read_data_block(rtems_device_minor_number  minor,
  */
 void
 ide_controller_write_data_block(rtems_device_minor_number  minor,
-                                uint16_t                   block_size,
+                                uint32_t                   block_size,
                                 rtems_blkdev_sg_buffer    *bufs,
                                 uint32_t                  *cbuf,
                                 uint32_t                  *pos)
